+++ paimon-e2e-tests/pom.xml
@@ -34,6 +34,7 @@ under the License.
     <properties>
         <flink.shaded.hadoop.version>2.8.3-10.0</flink.shaded.hadoop.version>
         <flink.sql.connector.kafka>flink-sql-connector-kafka</flink.sql.connector.kafka>
+        <flink.sql.connector.hive>flink-sql-connector-hive-2.3.9_${scala.binary.version}</flink.sql.connector.hive>
     </properties>
 
     <dependencies>
@@ -42,13 +43,6 @@ under the License.
         <dependency>
             <groupId>org.apache.paimon</groupId>
             <artifactId>paimon-flink-${test.flink.main.version}</artifactId>
-            <version>${project.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.paimon</groupId>
-            <artifactId>paimon-hive-catalog</artifactId>
             <version>${project.version}</version>
             <scope>runtime</scope>
         </dependency>
@@ -112,16 +106,6 @@ under the License.
                         </artifactItem>
                         <artifactItem>
                             <groupId>org.apache.paimon</groupId>
-                            <artifactId>paimon-hive-catalog</artifactId>
-                            <version>${project.version}</version>
-                            <destFileName>paimon-hive-catalog.jar</destFileName>
-                            <type>jar</type>
-                            <overWrite>true</overWrite>
-                            <outputDirectory>/tmp/paimon-e2e-tests-jars
-                            </outputDirectory>
-                        </artifactItem>
-                        <artifactItem>
-                            <groupId>org.apache.paimon</groupId>
                             <artifactId>paimon-hive-connector-common</artifactId>
                             <version>${project.version}</version>
                             <destFileName>paimon-hive-connector.jar</destFileName>
@@ -161,6 +145,17 @@ under the License.
                             <outputDirectory>/tmp/paimon-e2e-tests-jars
                             </outputDirectory>
                         </artifactItem>
+                        <!-- flink hive catalog needs flink hive sql jar -->
+                        <artifactItem>
+                            <groupId>org.apache.flink</groupId>
+                            <artifactId>${flink.sql.connector.hive}</artifactId>
+                            <version>${test.flink.version}</version>
+                            <destFileName>flink-sql-connector-hive.jar</destFileName>
+                            <type>jar</type>
+                            <overWrite>true</overWrite>
+                            <outputDirectory>/tmp/paimon-e2e-tests-jars
+                            </outputDirectory>
+                        </artifactItem>
                     </artifactItems>
                 </configuration>
             </plugin>
@@ -185,6 +180,7 @@ under the License.
                 <test.flink.main.version>1.14</test.flink.main.version>
                 <test.flink.version>1.14.6</test.flink.version>
                 <flink.sql.connector.kafka>flink-sql-connector-kafka_${scala.binary.version}</flink.sql.connector.kafka>
+                <flink.sql.connector.hive>flink-sql-connector-hive-2.3.6_${scala.binary.version}</flink.sql.connector.hive>
             </properties>
         </profile>
 
@@ -194,6 +190,7 @@ under the License.
                 <test.flink.main.version>1.15</test.flink.main.version>
                 <test.flink.version>1.15.3</test.flink.version>
                 <flink.sql.connector.kafka>flink-sql-connector-kafka</flink.sql.connector.kafka>
+                <flink.sql.connector.hive>flink-sql-connector-hive-2.3.6_${scala.binary.version}</flink.sql.connector.hive>
             </properties>
         </profile>
     </profiles>
