+++ paimon-hive/paimon-hive-connector-common/src/test/java/org/apache/paimon/hive/HiveCatalogITCaseBase.java
@@ -24,22 +24,17 @@ import org.apache.paimon.catalog.CatalogLock;
 import org.apache.paimon.catalog.Identifier;
 import org.apache.paimon.flink.FlinkCatalog;
 import org.apache.paimon.fs.local.LocalFileIO;
-import org.apache.paimon.schema.SchemaManager;
 
 import com.klarna.hiverunner.HiveShell;
-import com.klarna.hiverunner.annotations.HiveRunnerSetup;
 import com.klarna.hiverunner.annotations.HiveSQL;
-import com.klarna.hiverunner.config.HiveRunnerConfig;
 import org.apache.flink.connectors.hive.FlinkEmbeddedHiveRunner;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.table.api.EnvironmentSettings;
 import org.apache.flink.table.api.TableEnvironment;
-import org.apache.flink.table.api.TableException;
 import org.apache.flink.table.api.internal.TableEnvironmentImpl;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.CloseableIterator;
 import org.apache.flink.util.ExceptionUtils;
-import org.apache.hadoop.hive.ql.lockmgr.DbTxnManager;
 import org.assertj.core.api.Assertions;
 import org.junit.After;
 import org.junit.Assert;
@@ -57,38 +52,20 @@ import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static org.apache.hadoop.hive.conf.HiveConf.ConfVars.HIVE_IN_TEST;
-import static org.apache.hadoop.hive.conf.HiveConf.ConfVars.HIVE_SUPPORT_CONCURRENCY;
-import static org.apache.hadoop.hive.conf.HiveConf.ConfVars.HIVE_TXN_MANAGER;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
-/** IT cases for {@link HiveCatalog}. */
+/** IT cases for using Paimon {@link HiveCatalog} together with Paimon Hive connector. */
 @RunWith(FlinkEmbeddedHiveRunner.class)
-public class HiveCatalogITCase {
+public abstract class HiveCatalogITCaseBase {
 
     @Rule public TemporaryFolder folder = new TemporaryFolder();
 
-    private String path;
-    private TableEnvironment tEnv;
+    protected String path;
+    protected TableEnvironment tEnv;
 
     @HiveSQL(files = {})
-    private static HiveShell hiveShell;
-
-    @HiveRunnerSetup
-    private static final HiveRunnerConfig CONFIG =
-            new HiveRunnerConfig() {
-                {
-                    // catalog lock needs txn manager
-                    // hive-3.x requires a proper txn manager to create ACID table
-                    getHiveConfSystemOverride()
-                            .put(HIVE_TXN_MANAGER.varname, DbTxnManager.class.getName());
-                    getHiveConfSystemOverride().put(HIVE_SUPPORT_CONCURRENCY.varname, "true");
-                    // tell TxnHandler to prepare txn DB
-                    getHiveConfSystemOverride().put(HIVE_IN_TEST.varname, "true");
-                }
-            };
+    protected static HiveShell hiveShell;
 
     @Before
     public void before() throws Exception {
@@ -529,76 +506,7 @@ public class HiveCatalogITCase {
         Assert.assertEquals("[]", tables.toString());
     }
 
-    @Test
-    public void testCreateExistTableInHive() throws Exception {
-        tEnv.executeSql(
-                String.join(
-                        "\n",
-                        "CREATE CATALOG my_hive_custom_client WITH (",
-                        "  'type' = 'paimon',",
-                        "  'metastore' = 'hive',",
-                        "  'uri' = '',",
-                        "  'warehouse' = '" + path + "',",
-                        "  'metastore.client.class' = '"
-                                + CreateFailHiveMetaStoreClient.class.getName()
-                                + "'",
-                        ")"));
-        tEnv.executeSql("USE CATALOG my_hive_custom_client");
-        assertThatThrownBy(
-                        () ->
-                                tEnv.executeSql(
-                                                "CREATE TABLE hive_table(a INT, b INT, c INT, d INT)")
-                                        .await())
-                .isInstanceOf(TableException.class)
-                .hasMessage(
-                        "Could not execute CreateTable in path `my_hive_custom_client`.`default`.`hive_table`");
-        assertTrue(
-                new SchemaManager(
-                                LocalFileIO.create(),
-                                new org.apache.paimon.fs.Path(path, "default.db/hive_table"))
-                        .listAllIds()
-                        .isEmpty());
-    }
-
-    @Test
-    public void testAlterTableFailedInHive() throws Exception {
-        tEnv.executeSql(
-                        String.join(
-                                "\n",
-                                "CREATE CATALOG my_alter_hive WITH (",
-                                "  'type' = 'paimon',",
-                                "  'metastore' = 'hive',",
-                                "  'uri' = '',",
-                                "  'warehouse' = '" + path + "',",
-                                "  'metastore.client.class' = '"
-                                        + AlterFailHiveMetaStoreClient.class.getName()
-                                        + "'",
-                                ")"))
-                .await();
-        tEnv.executeSql("USE CATALOG my_alter_hive").await();
-        tEnv.executeSql("CREATE TABLE alter_failed_table(a INT, b STRING)").await();
-
-        assertThatThrownBy(() -> tEnv.executeSql("ALTER TABLE alter_failed_table SET ('aa'='bb')"))
-                .isInstanceOf(TableException.class)
-                .hasMessage(
-                        String.format(
-                                "Could not execute "
-                                        + "ALTER TABLE my_alter_hive.default.alter_failed_table "
-                                        + "SET (aa: [bb], path: [%sdefault.db/alter_failed_table])",
-                                path));
-
-        assertTrue(
-                new SchemaManager(
-                                LocalFileIO.create(),
-                                new org.apache.paimon.fs.Path(
-                                        path, "default.db/alter_failed_table"))
-                        .latest()
-                        .get()
-                        .options()
-                        .isEmpty());
-    }
-
-    private List<Row> collect(String sql) throws Exception {
+    protected List<Row> collect(String sql) throws Exception {
         List<Row> result = new ArrayList<>();
         try (CloseableIterator<Row> it = tEnv.executeSql(sql).collect()) {
             while (it.hasNext()) {
